% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_delay.R
\name{api_delay}
\alias{api_delay}
\title{API Delay Function}
\usage{
api_delay(start_time, wait_time = 1)
}
\arguments{
\item{start_time}{Time of last API call}

\item{wait_time}{Wait time between API calls}
}
\value{
No return. The function put the system to sleep if the time between
your last API call to the current one is less than 1s.
}
\description{
A function to control the wait time between API calls to OpenDota,
user needs to specify the start time of a call to the API. The wait time is 1s
by default since OpenDota ask you to restrict your call to 1 call per second
with a free API.
}
\examples{
\dontrun{
api_key(start_time, wait_time)
}
}
