% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{damred}
\alias{damred}
\title{The damage reduction provided by total armour (between zero and one).}
\usage{
damred(armour)
}
\arguments{
\item{armour}{input numeric vector}
}
\value{
numeric vector where 1 is total damage reduction and zero is none.
Can provide a negative which suggests negative armour value can give a negative damage reduction
i.e. the target takes additional damage
}
\description{
Note the hard-coded 0.06 multiplier which is a constant of the Dota universe.
}
\examples{
damred(1:5)
damred((0 - 7 - 20 - 10))  #  Natural order, desolator, amplify damage and solar crest; ouch!
plot(1:40, damred(1:40))
}
\seealso{
\code{\link{hp_multiplier}}
}

