% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{ehp}
\alias{ehp}
\title{Calculate the effective health points of a hero.}
\usage{
ehp(hp, armour)
}
\arguments{
\item{hp}{input numeric vector}

\item{armour}{input numeric vector}
}
\value{
numeric vector of effective hp
}
\description{
Effective HP is the amount of physical damage one can take before dying, and is based on the unit's HP.
Calculates effective hp by multiplying hp.multiplier (armour considered) by current hp.
Where effective hp means how much damage can a unit take before dying, after reductions.
In this case, armour increases a unit's effective hp against physical damage.
}
\examples{
ehp(625, 0)
ehp(3000, c(6, 16))  #  What happens if I buy a platemail?
}

