% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kill_log_df.R
\name{get_kill_log_df}
\alias{get_kill_log_df}
\title{A function to extract the kill log}
\usage{
get_kill_log_df(player_data)
}
\arguments{
\item{player_data}{A list returned by \code{get_player_data}, list}
}
\value{
The function that returns a dataframe containing player ID,
match ID, match date, in game time in second, name of enemy
hero killed.
}
\description{
A function to extract enemy heroes killed and the time
at which they were killed in a match
}
\examples{
\dontrun{
get_kill_log_df(player_data)
}
}
