% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lh_dn_df.R
\name{get_lh_dn_df}
\alias{get_lh_dn_df}
\title{A function to extract the last hit/deny data}
\usage{
get_lh_dn_df(player_data)
}
\arguments{
\item{player_data}{A list returned by \code{get_player_data}, lsit}
}
\value{
The function returns a dataframe containing player ID,
match ID, match date, last hit count, and deny count
}
\description{
A function to extract the last hit/deny data from
the list returned by \code{get_player_data}
}
\examples{
\dontrun{
get_lh_dn_df(player_data)
}
}
