% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_id.R
\name{get_match_id}
\alias{get_match_id}
\title{A function to get a numeric vector of \emph{n} match IDs}
\usage{
get_match_id(player_id, n)
}
\arguments{
\item{player_id}{The numeric ID of your Dota 2 account, numeric}

\item{n}{The number of match IDs you want, numeric}
}
\value{
The function returns a numeric vector containing the match IDs of the \emph{n}
most recent matchs that involves the player with the specific player ID provided by the user.
}
\description{
A function to fetch the match IDs of the \emph{n} most recent matchs
that involves the player with the specific player ID provided by the user.
}
\examples{
\dontrun{
get_match_id(123456, 20)
}
}
