% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_player_data}
\alias{get_player_data}
\title{A function to fetch the match data of a specific player}
\usage{
get_player_data(player_id, match_vec, wait_time = 1, summary_info = TRUE)
}
\arguments{
\item{player_id}{The numeric ID of your Dota 2 account, numeric}

\item{match_vec}{A numeric vector of at least one match ID, numeric}

\item{wait_time}{Wait time between API calls, numeric, default to 1s}

\item{summary_info}{Set to TRUE if you want summary information about
how many of the requested matchs are parsed by OpenDota, how many error
occured when calling the API, etc.}
}
\value{
The function returns a nested list. Each element in the list
is a parsed JSON file of a match.
}
\description{
A function to fetch the match data of a specific player
}
\examples{
\dontrun{
get_player_data(123456, 123456)
}

\dontrun{
get_player_data(123456, c(123456, 78910))
}

}
