% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purchase_log_df.R
\name{get_purchase_log_df}
\alias{get_purchase_log_df}
\title{A function to extract the purchase log}
\usage{
get_purchase_log_df(player_data)
}
\arguments{
\item{player_data}{A list returned by \code{get_player_data}}
}
\value{
The function returns a dataframe containing player ID,
match ID, match date, in game time in second, name of item purchased.
}
\description{
A function to extract item purchased and the time at which the item
was purchased
}
\examples{
\dontrun{
get_purchase_log_df(player_data)
}
}
