% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rune_log_df.R
\name{get_rune_log_df}
\alias{get_rune_log_df}
\title{A function to extract the rune log}
\usage{
get_rune_log_df(player_data)
}
\arguments{
\item{player_data}{A list returned by \code{get_player_data}}
}
\value{
The function returns a dataframe containing player ID,
match ID, match date, in game time in second, name of the rune you got.
}
\description{
A function to extract the type of rune a player got and the time
at which the rune was got.
}
\examples{
\dontrun{
get_rune_log_df(player_data)
}
}
