% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_win_lose_df.R
\name{get_win_lose_df}
\alias{get_win_lose_df}
\title{A function to extract win/lose status}
\usage{
get_win_lose_df(player_data)
}
\arguments{
\item{player_data}{A list returned by \code{get_player_data}}
}
\value{
The function returns a dataframe containing player ID,
match ID, match date, and the match outcome. Win = 1, loss = 0.
}
\description{
A function to extract the match outcome.
}
\examples{
\dontrun{
get_win_lose_df(player_data)
}
}
