% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{hp_multiplier}
\alias{hp_multiplier}
\title{The hit points multiplier is an armour adjusted factor which affects a heroes effective hp.}
\usage{
hp_multiplier(armour)
}
\arguments{
\item{armour}{input numeric vector}
}
\value{
numeric vector of hp multiplier to get effective hp
Can provide a negative which suggests negative armour value can give a negative damage reduction
i.e. the target takes additional damage
}
\description{
Calculates the hp multiplier, the
factor by which to times your current hp to get your effective hp.
Where effective hp means how much damage can a unit take before dying, after reductions.
In this case, armour increases a unit's effective hp against physical damage.
This is more intuitive to damage reduction as one can compare with the damage output offered by an enemy.
}
\examples{
hp.multiplier(-30)
}

