% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{mag_dam}
\alias{mag_dam}
\title{Magic damage to health points of a hero}
\usage{
mag_dam(magic_damage, mag_res = 0.35)
}
\arguments{
\item{mag_res}{magic resistance as a numeric vector}

\item{hp}{input numeric vector}
}
\value{
numeric vector of magic damage inflicted on target
}
\description{
Calculate the magic damage against a hero caused by a particular attack. Dependent on spell or attack damage and
the magic resistance of a hero which has a base of 0.35 but can vary due to auras and items.
Visage and Meepo have different base magic resistance (0.10 and 0.45, consult Dota wiki).
}
\examples{
mag_dam(600)  #  Threshold for Lion's Finger of Death Level 1 and base magic resist for most heroes.
sum(mag_dam(rep(40, 4)))  #  A strong level 1 nuke; Disruptor thunder strike
}

