% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{mag_ehp}
\alias{mag_ehp}
\title{Effective health points against magic damage}
\usage{
mag_ehp(hp, mag_res = 0.35)
}
\arguments{
\item{hp}{input numeric vector}

\item{mag_res}{magic resistance as a numeric vector}
}
\value{
numeric vector of effective hp against magic damage
}
\description{
Calculate the health pool available against magic damage of a hero. Dependent on hp and
the magic resistance of a hero which has a base of 0.35 but can vary due to auras and items.
Visage and Meepo have different base magic resistance (0.10 and 0.45, consult Dota wiki).
}
\examples{
mag_ehp(390)  #  Threshold for Lion's Finger of Death Level 1 and base magic resist for most heroes.
mag_ehp(200)  #  The minimum hp pool in Dota2, think decay stacks.
}

