% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dota_ehp.R
\name{phys_dam}
\alias{phys_dam}
\title{Physical damage to health points of a hero}
\usage{
phys_dam(physical_damage, armour)
}
\arguments{
\item{physical_damage}{input numeric vector}

\item{armour}{physical resistance as a numeric vector}
}
\value{
numeric vector of magic damage inflicted on target
}
\description{
Calculate the physical damage against a hero caused by a particular attack. Dependent on spell or attack damage and
the physical resistance which is based on the hero's armour.
}
\examples{
phys_dam(500, 0)  #  A level 1 Suicide Squad from Techies versus zero armour hero.
sum(phys_dam(rep(40, 2)))  #  Germinate attack from Weaver.
}

